STUDENTS_REPORT/requirements.txt
atom==0.4.1
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.13
django-allauth==0.34.0
django-rest-auth==0.9.2
django-rest-swagger==2.1.2
djangorestframework==3.7.7
future==0.16.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
MySQL-python==1.2.5
oauthlib==2.0.6
olefile==0.44
openapi-codec==1.3.2
Pillow==4.3.0
python-openid==2.2.5
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
